Fix bugs in detecting unchanging values.
The old code would fail if a new sysctl node appeared in the stream
of sysctl data because it would access a nonexistent key.  Detect
new nodes and don't die on key errors.

While here remove a debug printf.

Also fix the code when/if we're not eliminating duplicate rows,
then bad stuff would happen because we would unconditionally overwrite
the list of mibs to output with an empty set since we never ran the
duplication removal code.