Fix: sleep time, isodate, netstat and nfsstat
Allow SLEEP_SEC to be overridden via the environment.

'export' the ISODATE variable so it gets picked up in subshells when
we use single quotes to spawn them.

Fix nfsstat, add "read" that is needed to parse the data.

Fix netstat on interfaces.  Put the code needed to properly CSVify
the output.

Allow the netstat code to sleep for a variable number of seconds.