Don't derive file and column names in R
Instead of deriving the csv filename, column, and graphics output names
from a single arg, specify them individually to allow customization and
to prevent errors.

Fix the graph_it_all.sh to use this to avoid passing ./column name
as a column which was causing graphs to be output without datetime
being picked up properly.