Don't add the same nav item more than once.
This can happen because the algorithm for determining which
nodes to group together can find non-adjacent nodes to group
togther in the case something like this:

kern.foo
kern.bar
kern.long.item.with.lots.of.nodes
kern.short

in this case we would wind up with multiple entries in the navbar
for "kern".

Fix is simple, just check for already inserted nodes before inserting
so we only get a single entry for "kern".